feat: add validation utilities for message and message part validation
